# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: C4 Admin API
  version: 1.0.0
  description: '管理端 API（模板/账号/分组/规则/日志/统计）。认证：Authorization: Bearer <admin_token>。'
servers:
  - url: /api/admin
...
paths:
...
  /users:
    get:
      operationId: GetUsers
      tags: [users]
      summary: 用户列表（platform_admin 专属；分页/筛选/排序）
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 20, maximum: 200, description: 每页条数（上限 200，超限裁剪到 200） } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
        - { name: email, in: query, schema: { type: string } }
        - { name: sort, in: query, schema: { type: string } }
        - { name: order, in: query, schema: { type: string, enum: [asc, desc] } }
      responses:
        '200':
          description: 用户列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserListResponse' }
        default:
          $ref: '#/components/responses/Error'
    post:
      operationId: PostUsers
      tags: [users]
      summary: 创建用户（platform_admin 专属；email 唯一/格式、密码 ≤72 字节）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserCreate' }
      responses:
        '200':
          description: 创建后的用户（PasswordHash 永不下发）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        default:
          $ref: '#/components/responses/Error'
  /users/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    put:
      operationId: PutUsersId
      tags: [users]
      summary: 更新用户（role/status/max_concurrency/balance；变更即时生效——Auth 快照刷新）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserUpdate' }
      responses:
        '200':
          description: 更新后的用户
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        default:
          $ref: '#/components/responses/Error'
  /users/{id}/groups:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      operationId: GetUsersIdGroups
      tags: [users]
      summary: 读取用户被授予的分组与各专属倍率（platform_admin；用户视角，与 /groups/{id}/assignments 对称）
      responses:
        '200':
          description: 该用户被授予的组 id 列表与各专属倍率
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserGroupsResponse' }
        default:
          $ref: '#/components/responses/Error'
    put:
      operationId: PutUsersIdGroups
      tags: [users]
      summary: 设置用户的授予分组（platform_admin；替换语义：未列出即撤销，空数组 = 清空）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserGroupsBody' }
      responses:
        '200':
          description: 已生效的授予组 id 列表与专属倍率
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserGroupsResponse' }
        default:
          $ref: '#/components/responses/Error'
...
  /api/user/groups:
    get:
      operationId: GetUserGroups
      tags: [user]
      summary: 可选组列表（public 全部 + 已授予 private；无可用路由的 upstreams 组不返回）
      responses:
        '200':
          description: 分组列表
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Group' }
        default:
          $ref: '#/components/responses/Error'
...
components:
  responses:
    Error:
      description: 错误响应
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    GroupVisibility:
      type: string
      enum: [public, private]
    GroupProtocolConvert:
      type: string
      enum: [auto, chat_to_resp, mess_to_resp, resp_to_mess, chat_to_mess]
      description: 分组级协议转换模式。auto = 自动协商（优先原生协议，再按固定优先级尝试转换）；其余值为手动方向。off 不在枚举内——不转换 = 空数组（见 protocol_convert 字段说明）
...
    UserGroupsBody:
      type: object
      required: [group_ids]
      properties:
        group_ids:
          type: array
          items: { type: integer, format: int64 }
          description: 替换语义：完整授予组列表（未列出即撤销；空数组 = 清空）
        multipliers:
          type: object
          additionalProperties:
            type: number
            format: double
            nullable: true
            minimum: 0
            maximum: 10
          description: 可选：group_id → 该用户在该组的专属价格倍率（正常值，1 = ×1，0 = 免费，上限 10 = ×10；API 边界与万分数换算）。仅对 group_ids 中列出的组生效；null = 清除为未设置（回退组倍率）；未列出的组沿用当前值
    UserGroupsResponse:
      type: object
      required: [group_ids]
      properties:
        group_ids:
          type: array
          items: { type: integer, format: int64 }
        multipliers:
          type: object
          additionalProperties:
            type: number
            format: double
            nullable: true
          description: 该用户当前各授予组的专属价格倍率（正常值；null/缺省 = 未设置 → 用组倍率）
...
    UserRole:
      type: string
      enum: [platform_admin, user]
    UserStatus:
      type: string
      enum: [active, disabled]
    User:
      type: object
      properties:
        ID: { type: integer, format: int64 }
        Email: { type: string }
        Role: { $ref: '#/components/schemas/UserRole' }
        Status: { $ref: '#/components/schemas/UserStatus' }
        MaxConcurrency: { type: integer }
        Balance: { type: number, format: double, description: "余额 USD（浮点；内部存储毫分——1 USD = 100,000 毫分，API 边界换算）" }
        CreatedAt: { type: string, format: date-time }
        UpdatedAt: { type: string, format: date-time }
...
    UserCreate:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
        role: { $ref: '#/components/schemas/UserRole' }
        status: { $ref: '#/components/schemas/UserStatus' }
        max_concurrency: { type: integer, description: 用户级在途上限；0 = 不限 }
        balance: { type: number, format: double, description: "余额 USD（浮点，≥ 0；1 USD = 100,000 毫分）" }
    UserUpdate:
      type: object
      properties:
        role: { $ref: '#/components/schemas/UserRole' }
        status: { $ref: '#/components/schemas/UserStatus' }
        max_concurrency: { type: integer }
        balance: { type: number, format: double, description: "余额 USD（浮点，≥ 0；1 USD = 100,000 毫分）" }
    UserListResponse:
      type: object
      required: [total, rows]
      properties:
        total: { type: integer, format: int64 }
        rows:
          type: array
          items: { $ref: '#/components/schemas/User' }
...
    Group:
      type: object
      properties:
        ID: { type: integer, format: int64 }
        Name: { type: string }
        Visibility: { $ref: '#/components/schemas/GroupVisibility' }
        RoutingMode: { $ref: '#/components/schemas/GroupRoutingMode' }
        PriceMultiplier: { type: number, format: double, description: 价格倍率（正常值，1 = ×1，0 = 免费，上限 10 = ×10；API 边界与万分数换算——存储 15000 ↔ 显示 1.5） }
        AllowedModels:
          type: array
          items: { type: string }
        ProtocolConvert:
          type: array
          items: { $ref: '#/components/schemas/GroupProtocolConvert' }
          description: 协议转换模式（auto = 自动协商；空数组 = off = 不转换；手动方向按客户端格式命中）
        CreatedAt: { type: string, format: date-time }
        UpdatedAt: { type: string, format: date-time }
        DeletedAt: { type: string, format: date-time, nullable: true, description: 软删除时间戳；null = 存活（列表/消费路径过滤已删；GET 单个可查已删项） }
    GroupRoutingMode:
      type: string
      enum: [accounts, upstreams]
      description: accounts 使用现有账号池；upstreams 使用本组上游成员
...